fix: keep ordered test classes whole, prune split fixtures, fix vale dashes
