+++ backend/src/test/java/com/keplerops/groundcontrol/integration/MigrationSmokeSchemaShapeOfAssetRelationTest.java
@@ -1,7 +1,6 @@
 package com.keplerops.groundcontrol.integration;
 
 import jakarta.persistence.EntityManager;
-import java.util.List;
 import javax.sql.DataSource;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,113 +15,6 @@ class MigrationSmokeSchemaShapeOfAssetRelationTest extends BaseIntegrationTest {
 
     @Autowired
     private EntityManager entityManager;
-
-    /**
-     * Tables the migration set must leave queryable.
-     *
-     * One list plus one assertion replaces the 96 identical
-     * {@code SELECT 1 FROM <table> LIMIT 1} probes this test used to carry
-     * (issue #1467). The list is the readable inventory of what is covered,
-     * and a missing table now names itself in the failure instead of the run
-     * stopping at the first one.
-     */
-    private static final List<String> AUDITED_TABLES = List.of(
-            "requirement",
-            "requirement_audit",
-            "requirement_relation_audit",
-            "revinfo",
-            "traceability_link",
-            "github_issue_sync",
-            "requirement_import",
-            "traceability_link_audit",
-            "identity_user",
-            "identity_user_audit",
-            "identity_group_audit",
-            "group_membership_audit",
-            "identity_role_audit",
-            "role_permission_assignment_audit",
-            "role_grant_audit",
-            "project_access_grant_audit",
-            "operational_asset",
-            "operational_asset_audit",
-            "asset_relation",
-            "asset_relation_audit",
-            "asset_link",
-            "asset_link_audit",
-            "asset_external_id",
-            "asset_external_id_audit",
-            "observation",
-            "observation_audit",
-            "risk_scenario",
-            "risk_scenario_audit",
-            "risk_scenario_link",
-            "risk_scenario_link_audit",
-            "control",
-            "control_audit",
-            "control_link",
-            "control_link_audit",
-            "github_pr_sync",
-            "verification_result",
-            "verification_result_audit",
-            "registered_plugin",
-            "pack_registry_entry",
-            "pack_registry_entry_audit",
-            "pack_install_record",
-            "pack_install_record_audit",
-            "trust_policy",
-            "trust_policy_audit",
-            "threat_model",
-            "threat_model_audit",
-            "threat_model_link",
-            "threat_model_link_audit",
-            "users",
-            "authorities",
-            "finding",
-            "finding_audit",
-            "finding_link",
-            "finding_link_audit",
-            "control_test",
-            "control_test_audit",
-            "test_case",
-            "test_case_audit",
-            "test_case_step",
-            "test_case_step_audit",
-            "test_case_gherkin",
-            "test_case_gherkin_audit",
-            "asset_subtype_schema",
-            "asset_subtype_schema_audit",
-            "test_case_folder",
-            "test_case_folder_audit",
-            "test_plan",
-            "test_plan_audit",
-            "test_suite",
-            "test_suite_audit",
-            "test_suite_member",
-            "test_suite_member_audit",
-            "test_suite_source_requirement",
-            "test_suite_source_requirement_audit",
-            "audit",
-            "audit_audit",
-            "audit_link",
-            "audit_link_audit",
-            "test_run",
-            "test_run_audit",
-            "test_run_tester_assignment",
-            "test_run_tester_assignment_audit",
-            "test_run_case_result",
-            "test_run_case_result_audit",
-            "test_run_step_result",
-            "test_run_step_result_audit",
-            "scoped_control_implementation",
-            "scoped_control_implementation_audit",
-            "risk_control_mapping",
-            "risk_control_mapping_audit",
-            "mapping_observation",
-            "mapping_evidence",
-            "mapping_evidence_audit",
-            "research_intake",
-            "research_intake_audit",
-            "mcp_tool_event");
 
     /**
      * GC-M012 column-existence probes (V069 / V070). A column-by-column SELECT
