+++ backend/src/test/java/com/keplerops/groundcontrol/unit/domain/testcases/service/TestRunServiceTest.java
@@ -58,8 +58,6 @@ class TestRunServiceTest {
     private static final UUID TC2_ID = UUID.fromString("00000000-0000-0000-0000-000000000902");
     private static final UUID CASE_RESULT_ID = UUID.fromString("00000000-0000-0000-0000-000000000a01");
     private static final UUID STEP_RESULT_ID = UUID.fromString("00000000-0000-0000-0000-000000000b01");
-    private static final UUID OTHER_RUN_ID = UUID.fromString("00000000-0000-0000-0000-000000000c01");
-    private static final UUID OTHER_CASE_RESULT_ID = UUID.fromString("00000000-0000-0000-0000-000000000c02");
 
     private TestRunRepository testRunRepository;
     private TestRunTesterAssignmentRepository testerAssignmentRepository;
